Completed Part 3 and Customized Somewhat
Created a backup from when my site started to vary in style quite a bit from the tutorial. Because my pictures are vertical, it did not look good to me to have all the text on the top half of the picture. Considering having the style switch in phone mode or having the pictures disappear